fix: handle edge case with same review sha after force-push (#122)
When the force-push doesn't rewrite the whole history, then it may
happen that the review SHA and head SHA are the same. This means that
the changed files can't be resolved, like when the force-push rewrites
the whole history, but the action doesn't recognize that it's
"force-push" scenario and doesn't remove any review.

Simply comparing the review SHA and head SHA can be used to catch this
scenario and then treat it as a force-push.