fix: remove filter reviews by publish timestamp (#387)
Filtering the reviews by `publishedAt` is wrong, because the
`committedDate` of commit is a timestamp from when the commit is
created, not pushed to Github. So if the user has the commit locally for
some time, and someone reviews the PR in the meanwhile, the review is
not removed when the commit is pushed (because the review has a newer
timestamp than the commit).

The review filtering shouldn't be needed anyway, so it should be ok to
remove it completely.